Add sse headers to remotefs, flyte config
