+++ flytekit/configuration/__init__.py
@@ -530,6 +530,7 @@ class S3Config(object):
     backoff: datetime.timedelta = datetime.timedelta(seconds=5)
     access_key_id: typing.Optional[str] = None
     secret_access_key: typing.Optional[str] = None
+    s3_additional_kwargs: typing.Optional[str] = None
 
     @classmethod
     def auto(cls, config_file: typing.Union[str, ConfigFile] = None) -> S3Config:
@@ -546,6 +547,7 @@ class S3Config(object):
         kwargs = set_if_exists(kwargs, "backoff", _internal.AWS.BACKOFF_SECONDS.read(config_file))
         kwargs = set_if_exists(kwargs, "access_key_id", _internal.AWS.S3_ACCESS_KEY_ID.read(config_file))
         kwargs = set_if_exists(kwargs, "secret_access_key", _internal.AWS.S3_SECRET_ACCESS_KEY.read(config_file))
+        kwargs = set_if_exists(kwargs, "s3_additional_kwargs", _internal.AWS.S3_ADDITIONAL_KWARGS.read(config_file))
         return S3Config(**kwargs)
