+++ flytekit/core/utils.py
@@ -211,6 +211,29 @@ def write_proto_to_file(proto, path):
         writer.write(proto.SerializeToString())
 
 
+def get_extra_headers_for_signed_url(signed_url):
+    """
+    A limitation of S3 Bucket Policies is that when enforcing SSE, checks are
+    done on the headers of the request. Therefore, when using pre-signed urls
+    we should fill in the headers with the SSE information from the url itself.
+
+    This is possible since these self-contained pre-signed urls contain the SSE
+    information within its query string.
+    """
+    from urllib.parse import parse_qs
+    headers = dict()
+    query_params = parse_qs(signed_url)
+    if "x-amz-server-side-encryption" in query_params:
+        sse = query_params["x-amz-server-side-encryption"]
+        if len(sse) > 0:
+            headers["x-amz-server-side-encryption"] = sse[0]
+    if "x-amz-server-side-encryption-aws-kms-key-id" in query_params:
+        keyid = query_params["x-amz-server-side-encryption-aws-kms-key-id"]
+        if len(keyid) > 0:
+            headers["x-amz-server-side-encryption-aws-kms-key-id"] = keyid[0]
+    return headers
+
+
 class Directory(object):
     def __init__(self, path):
         """
